pyproject.toml
[tool.poetry]
name = "examstutor-ai-api"
version = "0.3.0"
description = "ExamsTutor AI Tutor API - Phase 3: Production Deployment with Offline Capabilities"
authors = ["ExamsTutor Team"]
readme = "README.md"
python = "^3.10"

[tool.poetry.dependencies]
python = "^3.10"
# API Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# ML & AI
torch = "^2.1.0"
transformers = "^4.36.0"
accelerate = "^0.25.0"
bitsandbytes = "^0.41.0"  # For quantization
peft = "^0.7.0"  # For LoRA/QLoRA
sentencepiece = "^0.1.99"

# Quantization & Optimization (Epic 3.1)
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
onnxruntime-gpu = "^1.16.3"
optimum = {extras = ["onnxruntime"], version = "^1.16.0"}
auto-gptq = "^0.6.0"
autoawq = "^0.1.8"

# Vector Database for Offline RAG
qdrant-client = "^1.7.0"
chromadb = "^0.4.22"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"

# Database
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"

# Caching & Background Tasks
redis = "^5.0.1"
celery = "^5.3.4"

# Monitoring & Observability (Epic 3.3)
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
python-json-logger = "^2.0.7"

# Testing (Epic 3.2)
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
locust = "^2.20.0"  # Load testing

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^41.0.0"
python-multipart = "^0.0.6"

# Utilities
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"  # Retry logic
loguru = "^0.7.2"
pyyaml = "^6.0.1"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.0"
pre-commit = "^3.6.0"
ipython = "^8.19.0"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
